+++ test/data/services/context_builder_service_test.mocks.dart
@@ -390,6 +390,25 @@ class MockPersonRepository extends _i1.Mock implements _i10.PersonRepository {
       ) as _i7.Future<void>);
 
   @override
+  _i7.Future<void> unlinkPartner(
+    String? userId,
+    String? personId,
+    String? partnerId,
+  ) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #unlinkPartner,
+          [
+            userId,
+            personId,
+            partnerId,
+          ],
+        ),
+        returnValue: _i7.Future<void>.value(),
+        returnValueForMissingStub: _i7.Future<void>.value(),
+      ) as _i7.Future<void>);
+
+  @override
   _i7.Future<void> deletePerson(
     String? userId,
     String? personId,