+++ test/presentation/screens/home_screen_test.mocks.dart
@@ -612,6 +612,25 @@ class MockPersonRepository extends _i1.Mock implements _i12.PersonRepository {
       ) as _i6.Future<void>);
 
   @override
+  _i6.Future<void> unlinkPartner(
+    String? userId,
+    String? personId,
+    String? partnerId,
+  ) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #unlinkPartner,
+          [
+            userId,
+            personId,
+            partnerId,
+          ],
+        ),
+        returnValue: _i6.Future<void>.value(),
+        returnValueForMissingStub: _i6.Future<void>.value(),
+      ) as _i6.Future<void>);
+
+  @override
   _i6.Future<void> deletePerson(
     String? userId,
     String? personId,